refactor 메소드 위치 수정 #188
