+++ src/main/java/poomasi/domain/farm/controller/FarmFarmerController.java
@@ -7,8 +7,8 @@ import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 import poomasi.domain.auth.security.userdetail.UserDetailsImpl;
-import poomasi.domain.farm.dto.FarmRegisterRequest;
-import poomasi.domain.farm.dto.FarmUpdateRequest;
+import poomasi.domain.farm.dto.request.FarmRegisterRequest;
+import poomasi.domain.farm.dto.request.FarmUpdateRequest;
 import poomasi.domain.farm.service.FarmFarmerService;
 import poomasi.domain.farm._schedule.service.FarmScheduleService;
 import poomasi.domain.member.entity.Member;
