+++ src/main/java/poomasi/domain/farm/dto/request/FarmRegisterRequest.java
@@ -1,4 +1,4 @@
-package poomasi.domain.farm.dto;
+package poomasi.domain.farm.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import poomasi.domain.farm.entity.Farm;
