+++ src/main/java/poomasi/domain/farm/service/FarmPlatformService.java
@@ -3,7 +3,7 @@ package poomasi.domain.farm.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import poomasi.domain.farm.dto.FarmResponse;
+import poomasi.domain.farm.dto.response.FarmResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
